@@ -16,16 +16,15 @@
         }
     ],
     "require": {
-        "php": "^8.1|^8.2",
-        "illuminate/contracts": "^10.0|^11.0|^12.0",
-        "illuminate/support": "^10.0|^11.0|^12.0",
+        "php": "^8.2",
+        "illuminate/contracts": "^11.0|^12.0",
+        "illuminate/support": "^11.0|^12.0",
         "spatie/laravel-package-tools": "^1.9"
     },
     "require-dev": {
         "guzzlehttp/guzzle": "^7.4",
         "laravel/pint": "^1.10",
-        "nunomaduro/collision": "^6.0|^8.1",
-        "orchestra/testbench": "^7.17|^8.0|^10.0",
+        "orchestra/testbench": "^9.0|^10.0",
         "phpstan/extension-installer": "^1.1",
         "phpstan/phpstan-deprecation-rules": "^1.0|^2.0",
         "phpstan/phpstan-phpunit": "^1.0|^2.0",
