@@ -28,64 +28,64 @@ class ImportmapTest extends TestCase
         $this->map->pinAllFrom('resources/js/libs', to: 'js/libs');
     }
 
-    /** @test */
-    public function local_bin_with_inferred_to()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function local_bin_with_inferred_to(): void
     {
         $this->assertEquals(asset('js/app.js'), Arr::get($this->map->asArray('asset'), 'imports.app'));
     }
 
-    /** @test */
-    public function local_pin_with_explicit_to()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function local_pin_with_explicit_to(): void
     {
         $this->assertEquals(asset('js/rich_text.js'), Arr::get($this->map->asArray('asset'), 'imports.editor'));
     }
 
-    /** @test */
-    public function remote_pin_works()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function remote_pin_works(): void
     {
         $this->assertEquals('https://cdn.skypack.dev/md5', Arr::get($this->map->asArray('asset'), 'imports.md5'));
     }
 
-    /** @test */
-    public function directory_pin_mounted_under_matching_subdir_maps_all_files()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function directory_pin_mounted_under_matching_subdir_maps_all_files(): void
     {
         $this->assertEquals(asset('js/controllers/hello_controller.js'), Arr::get($this->map->asArray('asset'), 'imports.controllers/hello_controller'));
         $this->assertEquals(asset('js/controllers/utilities/md5_controller.js'), Arr::get($this->map->asArray('asset'), 'imports.controllers/utilities/md5_controller'));
     }
 
-    /** @test */
-    public function vendor_directory_inside_pinned_folder_is_ignored()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function vendor_directory_inside_pinned_folder_is_ignored(): void
     {
         $this->assertNull(Arr::get($this->map->asArray('asset'), 'imports.vendor/alpine'));
     }
 
-    /** @test */
-    public function directory_pin_mounted_under_matching_subdir_maps_index_as_root()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function directory_pin_mounted_under_matching_subdir_maps_index_as_root(): void
     {
         $this->assertEquals(asset('js/controllers/index.js'), Arr::get($this->map->asArray('asset'), 'imports.controllers'));
     }
 
-    /** @test */
-    public function directory_pin_mounted_under_matching_subdir_maps_index_as_root_at_second_depth()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function directory_pin_mounted_under_matching_subdir_maps_index_as_root_at_second_depth(): void
     {
         $this->assertEquals(asset('js/helpers/requests/index.js'), Arr::get($this->map->asArray('asset'), 'imports.helpers/requests'));
     }
 
-    /** @test */
-    public function directory_pin_under_custom_asset_path()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function directory_pin_under_custom_asset_path(): void
     {
         $this->assertEquals(asset('js/spina/controllers/another_controller.js'), Arr::get($this->map->asArray('asset'), 'imports.controllers/spina/another_controller'));
         $this->assertEquals(asset('js/spina/controllers/deeper/again_controller.js'), Arr::get($this->map->asArray('asset'), 'imports.controllers/spina/deeper/again_controller'));
     }
 
-    /** @test */
-    public function directory_pin_without_path_or_under()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function directory_pin_without_path_or_under(): void
     {
         $this->assertEquals(asset('my_lib.js'), Arr::get($this->map->asArray('asset'), 'imports.my_lib'));
     }
 
-    /** @test */
-    public function preload_modules_are_included_in_preload_tags()
+    #[\PHPUnit\Framework\Attributes\Test]
+    public function preload_modules_are_included_in_preload_tags(): void
     {
         $preloadingModulePaths = json_encode($this->map->preloadedModulePaths('asset'));
 
